checkpointer: factor out checkpoint lib.
The monolithic main.go file was becoming too unwieldy to keep working
on. This is a straightforward refactor: I moved functions into logical
files and made the main() as lightweight as possible.
